+++ packages/compose/src/index.native.js
@@ -17,7 +17,7 @@ export { default as withInstanceId } from './higher-order/with-instance-id';
 export { default as withSafeTimeout } from './higher-order/with-safe-timeout';
 export { default as withState } from './higher-order/with-state';
 export { default as withPreferredColorScheme } from './higher-order/with-preferred-color-scheme';
-export { default as withIsConnected } from './higher-order/with-is-connected';
+export { default as withNetworkConnectivity } from './higher-order/with-network-connectivity';
 
 // Hooks.
 export { default as useConstrainedTabbing } from './hooks/use-constrained-tabbing';
@@ -38,3 +38,4 @@ export { default as useDebouncedInput } from './hooks/use-debounced-input';
 export { default as useThrottle } from './hooks/use-throttle';
 export { default as useMergeRefs } from './hooks/use-merge-refs';
 export { default as useRefEffect } from './hooks/use-ref-effect';
+export { default as useNetworkConnectivity } from './hooks/use-network-connectivity';