+++ phpunit/tests/fonts/font-library/wpRestFontCollectionsController/base.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Test Case for WP_REST_Font_Library_Controller tests.
+ * Test Case for WP_REST_Font_Collections_Controller tests.
  *
  * @package WordPress
  * @subpackage Font Library
  */
-abstract class WP_REST_Font_Library_Controller_UnitTestCase extends WP_UnitTestCase {
+abstract class WP_REST_Font_Collections_Controller_UnitTestCase extends WP_UnitTestCase {
 
 	/**
 	 * Fonts directory.
@@ -17,8 +17,6 @@ abstract class WP_REST_Font_Library_Controller_UnitTestCase extends WP_UnitTestC
 
 	public function set_up() {
 		parent::set_up();
-
-		static::$fonts_dir = WP_Font_Library::get_fonts_dir();
 
 		// Create a user with administrator role.
 		$admin_id = $this->factory->user->create(
@@ -40,10 +38,5 @@ abstract class WP_REST_Font_Library_Controller_UnitTestCase extends WP_UnitTestC
 		$property   = $reflection->getProperty( 'collections' );
 		$property->setAccessible( true );
 		$property->setValue( null, array() );
-
-		// Clean up the /fonts directory.
-		foreach ( $this->files_in_dir( static::$fonts_dir ) as $file ) {
-			@unlink( $file );
-		}
 	}
 }